chore: move panoramapopup and psv.js to folder
